requirements.txt
pandas==2.2.0
icalendar==5.0.11
python-dotenv==1.0.1
requests==2.31.0
psycopg2-binary==2.9.9
fastapi==0.116.1
uvicorn==0.35.0
python-multipart==0.0.20
email-validator==2.2.0
